requirements.txt
Flask==3.1.0
flask-restx==1.3.0
python-dotenv==1.0.1
requests-oauthlib==2.0.0